\name{f.analyzeFMRI.gui}
\alias{f.analyzeFMRI.gui}
\title{starts AnalyzeFMRI GUI}
\description{Starts an R/tk interfaced GUI that allows the user to
  explore an fMRI dataset stored in an ANALYZE format file using the
  functions of the AnalyzeFMRI package.
}
\usage{
f.analyzeFMRI.gui()
}
\value{No value is returned
}
\keyword{utilities}

