\name{f.plot.volume.gui}
\alias{f.plot.volume.gui}
\title{tcltk GUI to display FMRI or MRI images}
\description{tcltk GUI to display FMRI or MRI images. This GUI is very
  usefull, for example, for investigating the results of an ICA
  performed with f.icast.fmri.gui(). But it can also be used to display
  an MRI or an FMRI image}
\usage{f.plot.volume.gui()}
\value{Nothing}
\details{
One has the possibility to enter either a filename (with its path)
or directly an R object in the file field.
}
\author{P Lafaye de Micheaux <plafaye@club.fr>}

\seealso{
  \code{\link{f.icast.fmri.gui}}
}
\examples{
# TODO!!
}
\keyword{utilities}



