\name{model.2.cov.func}
\alias{model.2.cov.func}
\title{Calculates covariance from Hartvig Model 2}
\description{Calculates covariance from Hartvig Model 2
}
\usage{
model.2.cov.func(g, par)
}
\arguments{
  \item{g}{The value of gamma}
  \item{par}{A vector of parameters from the N2G model}
}
\value{The calculated covariance
}
\references{Hartvig, N. V. and Jensen, J. L (2000) Spatial Mixture
  Modelling of fMRI Data, Human Brain Mapping 11:233--248}
\author{J. L. Marchini}

\keyword{utilities}
