\name{model.2.est.gamma}
\alias{model.2.est.gamma}
\title{Estimate gamma for Model 2 of Hartvig and Jensen (2000)}
\description{Estimate gamma for Model 2 of Hartvig and Jensen (2000)
}
\usage{
model.2.est.gamma(cov, par)
}
\arguments{
  \item{cov}{An estimate of the spatial covariance}
  \item{par}{N2G model parameter estimates}
}
\value{The estimate of gamma
}
\references{Hartvig, N. V. and Jensen, J. L (2000) Spatial Mixture
  Modelling of fMRI Data, Human Brain Mapping 11:233--248}
\author{J. L. Marchini}

\keyword{utilities}
