\name{N2G.Class.Probability}
\alias{N2G.Class.Probability}
\title{Posterior Probabilities for N2G model}
\description{Calculates the Posterior Probability of data points being
  in each class given the parameters of the N2G model.
}
\usage{
N2G.Class.Probability(data, par)
}
\arguments{
  \item{data}{The dataset (usually a vector)}
  \item{par}{The paarmeters of the model}
}
\value{Returns the Posterior Probability of data points being
  in each class given the parameters of the N2G model.}
\author{J. L. Marchini}
\seealso{\code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
  \code{\link{N2G.Density}}, \code{\link{N2G.Likelihood}} ,   \code{\link{N2G.Transform}},
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
 \code{\link{N2G.Inverse}}   ,         \code{\link{N2G.Region}}
}
\keyword{utilities}





