\name{N2G.Density}
\alias{N2G.Density}

\title{Calculates the density function for the N2G model}
\description{Calculates the density function for the N2G model
 
}
\usage{
N2G.Density(data, par)
}

\arguments{
  \item{data}{The dataset (usually a vector)}
  \item{par}{The parameters of the model.}
}
\details{Calculates the density function for the N2G model
}
\value{Returns the density at each point of the datasets
 
}
\author{J. L. Marchini}


\seealso{\code{\link{N2G.Class.Probability}},  \code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
         \code{\link{N2G.Likelihood}} ,   \code{\link{N2G.Transform}},
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
 \code{\link{N2G.Inverse}}   ,         \code{\link{N2G.Region}}
}



\keyword{utilities}
