\name{N2G.Inverse}
\alias{N2G.Inverse}
\title{Transform parameters of N2G model back to their real domains}
\description{Transform parameters of N2G model back to their real domains
}
\usage{
N2G.Inverse(par)
}
\arguments{
  \item{par}{Parameter vector}
}
\details{Transform parameters of N2G model back to their real domains
}
\value{Returns the transformed parameters.
}
\author{J. L. Marchini}

\seealso{\code{\link{N2G.Class.Probability}},  \code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
 \code{\link{N2G.Density}} ,         \code{\link{N2G.Likelihood}} ,   \code{\link{N2G.Transform}},
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
         \code{\link{N2G.Region}}
}


\keyword{utilities}
