\name{N2G.Likelihood}
\alias{N2G.Likelihood}
\title{Calculates the (negative) Likelihood of the N2G model}
\description{Calculates the (negative) Likelihood of the N2G model
}
\usage{
N2G.Likelihood(inv.par, data)
}
\arguments{
  \item{inv.par}{A vector of transformed parameters for the N2G model}
  \item{data}{The dataset (usually a vector)}
}
\details{Calculates the (negative) Likelihood of the N2G model
}
\value{Returns (negative) Likelihood at each point of the dataset.
}
\author{J. L. Marchini}


\seealso{\code{\link{N2G.Class.Probability}},  \code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
 \code{\link{N2G.Density}} ,       \code{\link{N2G.Transform}},
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
 \code{\link{N2G.Inverse}}   ,         \code{\link{N2G.Region}}
}

\keyword{utilities}
