\name{N2G.Region}
\alias{N2G.Region}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{N2G Normal component interval}
\description{Calculates the interval within which observations are
  classified as belonging to the Normal component of an N2G model.
}
\usage{
N2G.Region(par1)
}
\arguments{
  \item{par1}{The parameters of the N2G model.}
}
\value{Avector containing the upper and lower boundaries of the interval.
}
\author{J. L. Marchini}

\seealso{\code{\link{N2G.Class.Probability}},  \code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
 \code{\link{N2G.Density}} ,         \code{\link{N2G.Likelihood}} ,   \code{\link{N2G.Transform}},
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
 \code{\link{N2G.Inverse}} 
}

\keyword{utilities}
