\name{N2G.Transform}
\alias{N2G.Transform}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Transform parameters of N2G model so as to lie on the real line.}
\description{Transform parameters of N2G model so as to lie on the real line
}
\usage{
N2G.Transform(par)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{par}{Parameter vector for N2G model.}
}
\details{Transformation required for optimization.
}
\value{Returns the transformed parameters.
}
\author{J. L. Marchini}

\seealso{\code{\link{N2G.Class.Probability}},  \code{\link{N2G.Likelihood.Ratio}},  \code{\link{N2G.Spatial.Mixture}},
 \code{\link{N2G.Density}} ,         \code{\link{N2G.Likelihood}} ,
 \code{\link{N2G.Fit}}    ,         \code{\link{N2G}} ,          
 \code{\link{N2G.Inverse}}   ,         \code{\link{N2G.Region}}
}

\keyword{utilities}
