\name{Q2R}
\alias{Q2R}
\title{Quaternion to rotation}
\description{Generate a rotation matrix from a quaternion.}
\usage{Q2R(Q,qfac)}
\arguments{\item{Q}{quaternion vector}
  \item{qfac}{qfac nifti field. It is pixdim[1]}
}
\value{The rotation.}
\seealso{
  \code{\link{R2Q}}
}
\examples{
}
\keyword{utilities}



